Update checked numerics library.
This library hadn't been updated in quite some time and had bugs
which affected UBSAN in some operators.

Bug: angleproject:5795
Reviewed-on: https://chromium-review.googlesource.com/c/angle/angle/+/3289941
Commit-Queue: Jamie Madill <jmadill@chromium.org>

+++ src/libANGLE/renderer/gl/renderergl_utils.cpp
@@ -2603,7 +2603,7 @@ angle::Result ShouldApplyLastRowPaddingWorkaround(ContextGL *contextGL,
     // At this point checkedEndByte is the actual last byte read.
     // The driver adds an extra row padding (if any), mimic it.
     ANGLE_CHECK_GL_MATH(contextGL, checkedPixelBytes.IsValid());
-    if (checkedPixelBytes.ValueOrDie() * size.width < rowPitch)
+    if (static_cast<size_t>(checkedPixelBytes.ValueOrDie()) * size.width < rowPitch)
     {
         checkedEndByte += rowPitch - checkedPixelBytes * size.width;
     }
